requirements-dev.txt
-r requirements.txt
black==22.1.0
click==8.0.4
flake8==4.0.1
isort==5.10.1
pytest==7.0.0
pytest-cov==3.0.0
requests-mock==1.9.3